requirements.txt
langgraph==0.2.39
langchain==0.3.7
langchain-openai==0.2.10
langchain-mcp-adapters==0.1.0
python-dotenv==1.0.1
# Added for the met-museum stdio transport if needed by the adapter
pydantic>=2.0.0
